+++ CHANGELOG.md
@@ -9,16 +9,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-* Adding unique constraint for categorisations
-* Adding organisation importer
+* Additional people ORCID mappings
+* Additional organisations for people
+* Allowing GTR importer instances to accept a GTR resource URI
+* Exceptions for unmapped GTR resources
+* Unique database constraint for categorisations
+* Organisation importer command
 * Integration tests for custom CLI commands
 
 ### Fixed
 
+* Support for unclassified grants in GTR importer
+* Difference in handling GTR project topics compared to categories in GTR importer
+* Application exception to act as a generic error class for all types of error, not just those at the API layer
 * Table name on categorisations migration statement
 
 ### Changed
 
+* Refactoring finding distinct GTR project category/topics in GTR importer
+* Refactoring API exception to inherit from new application exception
 * Refactoring importers into a package
 * Refactoring commands into main application
 * Refactoring standalone routes into a separate module